Session5/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
tqdm>=4.62.0
Pillow>=8.3.1
numpy>=1.21.0
torchsummary>=1.5.1
matplotlib>=3.4.0
pytest>=7.0.0
pytest-cov>=3.0.0 